Added Reddit About wrapper;

+++ Nicruo.ReddSharp/Domain/Post.cs
@@ -10,9 +10,9 @@ namespace Nicruo.ReddSharp.Domain
     {
         public string Title { get; set; }
         public string Author { get; set; }
-        public int Score { get; set; }
+        public long Score { get; set; }
         public string Thumbnail { get; set; }
         public string Url { get; set; }
-        public int NumberOfComments { get; set; }
+        public long NumberOfComments { get; set; }
     }
 }
